Add missing getChannel() methods in javalib
This also brought a minor refactor, specifically to FileChannelImpl and
RandomAccessFile. Since previously FCI used RAF, and RAF was supposed
to use FCI, some methods were moved to FCI, though they are stil
callable in RAF, FileInputStream and FileOutputStream the same way as
they were before. Tests testing the new functionality were also added.
