+++ javalib/src/main/scala/java/io/FileOutputStream.scala
@@ -13,6 +13,7 @@ import scala.scalanative.windows.FileApi._
 import scala.scalanative.windows.FileApiExt._
 import scala.scalanative.windows.HandleApiExt._
 import scala.scalanative.windows.winnt.AccessRights._
+import java.nio.channels.{FileChannelImpl, FileChannel}
 
 class FileOutputStream(fd: FileDescriptor, file: Option[File])
     extends OutputStream {
@@ -22,6 +23,9 @@ class FileOutputStream(fd: FileDescriptor, file: Option[File])
   def this(file: File) = this(file, false)
   def this(name: String, append: Boolean) = this(new File(name), append)
   def this(name: String) = this(new File(name))
+
+  private val channel: FileChannelImpl =
+    new FileChannelImpl(fd, file, deleteOnClose = false)
 
   override def close(): Unit = fd.close()
 
@@ -37,43 +41,13 @@ class FileOutputStream(fd: FileDescriptor, file: Option[File])
     write(buffer, 0, buffer.length)
   }
 
-  override def write(buffer: Array[Byte], offset: Int, count: Int): Unit = {
-    if (buffer == null) {
-      throw new NullPointerException
-    }
-    if (offset < 0 || count < 0 || count > buffer.length - offset) {
-      throw new IndexOutOfBoundsException
-    }
-    if (count == 0) {
-      return
-    }
-
-    // we use the runtime knowledge of the array layout to avoid
-    // intermediate buffer, and read straight from the array memory
-    val buf = buffer.asInstanceOf[runtime.ByteArray].at(offset)
-    if (isWindows) {
-      val hasSucceded =
-        WriteFile(fd.handle, buf, count.toUInt, null, null)
-      if (!hasSucceded) {
-        throw WindowsException.onPath(
-          file.fold("<file descriptor>")(_.toString)
-        )
-      }
-    } else {
-      val writeCount = unistd.write(fd.fd, buf, count.toUInt)
-
-      if (writeCount < 0) {
-        // negative value (typically -1) indicates that write failed
-        throw UnixException(file.fold("")(_.toString), errno.errno)
-      }
-    }
-  }
+  override def write(buffer: Array[Byte], offset: Int, count: Int): Unit =
+    channel.write(buffer, offset, count)
 
   override def write(b: Int): Unit =
     write(Array(b.toByte))
 
-  // TODO:
-  // def getChannel(): FileChannel
+  def getChannel(): FileChannel = channel
 }
 
 object FileOutputStream {
