+++ javalib/src/main/scala/java/io/RandomAccessFile.scala
@@ -9,6 +9,7 @@ import scalanative.meta.LinktimeInfo.isWindows
 import scala.scalanative.windows
 import windows._
 import windows.FileApiExt._
+import java.nio.channels.{FileChannelImpl, FileChannel}
 
 class RandomAccessFile private (
     file: File,
@@ -30,13 +31,16 @@ class RandomAccessFile private (
   private var closed: Boolean = false
   private lazy val in = new DataInputStream(new FileInputStream(fd))
   private lazy val out = new DataOutputStream(new FileOutputStream(fd))
+  private lazy val channel =
+    new FileChannelImpl(fd, Some(file), deleteOnClose = false)
 
   override def close(): Unit = {
     closed = true
     if (isWindows) HandleApi.CloseHandle(fd.handle)
     else unistd.close(fd.fd)
   }
-  // final def getChannel(): FileChannel
+
+  final def getChannel(): FileChannel = channel
 
   def getFD(): FileDescriptor =
     fd
